Safely handle missing credentials.d folder
The `$HOME/.aws/credentials.d` folder is not required to exist, as all
credential configurations can be defined on `$HOME/.aws/credentials` instead.

The early return with the `?` operator caused the binary to fail if the
optional folder did not exist, but worked if the folder existed but was empty.

This commit refactors the code the avoid crashing when the
`$HOME/.aws/credentials.d` does not exist.

It is collecting into a Vector instead of chaining the streams as it would
required Boxed Streams to switch between `ReadDirStream` and `stream::empty` in
case of an `Err`.